+++ ufo/config/config.yaml.template
@@ -1,9 +1,9 @@
 version: 0.2
 
 API_TYPE: "azure_ad"  # The API type, "openai" for the OpenAI API, "aoai" for the AOAI API, 'azure_ad' for the ad authority of the AOAI API.
-OPENAI_API_BASE: "https://cloudgpt-openai.azure-api.net/" # The the OpenAI API endpoint, "https://api.openai.com/v1/chat/completions" for the OpenAI API. As for the AAD, it should be your endpoints.
+OPENAI_API_BASE: "YOUR_ENDPOINT" # The the OpenAI API endpoint, "https://api.openai.com/v1/chat/completions" for the OpenAI API. As for the AAD, it should be your endpoints.
 OPENAI_API_KEY: "YOUR_API_KEY"  # The OpenAI API key
-API_VERSION: "2023-12-01-preview" 
+API_VERSION: "API_VERSION" # 
 OPENAI_API_MODEL: "gpt-4-visual-preview"  # The only OpenAI model by now that accepts visual input
 CONTROL_BACKEND: "uia"  # The backend for control action
 MAX_TOKENS: 2000  # The max token limit for the response completion
@@ -37,9 +37,9 @@ APP_SELECTION_PROMPT: "ufo/prompts/base/app_selection.yaml"  # The prompt for th
 ACTION_SELECTION_PROMPT: "ufo/prompts/base/action_selection.yaml"  # The prompt for the action selection
 INPUT_TEXT_API: "type_keys" # The input text API
 ###For AAD
-AAD_TENANT_ID: "72f988bf-86f1-41af-91ab-2d7cd011db47"
-AAD_API_SCOPE: "openai" #"openai"
-AAD_API_SCOPE_BASE: "feb7b661-cac7-44a8-8dc1-163b63c23df2" #API://YOUR_SCOPE_BASE
+AAD_TENANT_ID: "YOUR_TENANT_ID"
+AAD_API_SCOPE: "YOUR_SCOPE" #"openai"
+AAD_API_SCOPE_BASE: "YOUR_SCOPE_BASE" #API://YOUR_SCOPE_BASE
