+++ ufo/module/flow.py
@@ -88,8 +88,9 @@ Please enter your request to be completed🛸: """.format(art=text2art("UFO")), 
             response, cost = llm_call.get_gptv_completion(app_selection_prompt_message, headers)
 
         except Exception as e:
+            print(e)
             log = json.dumps({"step": self.step, "status": str(e), "prompt": app_selection_prompt_message})
-            print_with_color("Error occurs when calling LLM2.", "red")
+            print_with_color("Error occurs when calling LLM.", "red")
             self.request_logger.info(log)
             self.status = "ERROR"
             return
@@ -98,7 +99,7 @@ Please enter your request to be completed🛸: """.format(art=text2art("UFO")), 
         self.cost += cost
 
         try:
-            aad = configs['API_TYPE'].lower() == 'azure'
+            aad = configs['API_TYPE'].lower() == 'azure_ad'
             if not aad:
                 response_string = response["choices"][0]["message"]["content"]
             else:
@@ -214,7 +215,7 @@ Please enter your request to be completed🛸: """.format(art=text2art("UFO")), 
                 response, cost = llm_call.get_gptv_completion(action_selection_prompt_message, headers)
             except Exception as e:
                 log = json.dumps({"step": self.step, "status": str(e), "prompt": action_selection_prompt_message})
-                print_with_color("Error occurs when calling LLM1.", "red")
+                print_with_color("Error occurs when calling LLM.", "red")
                 self.request_logger.info(log)
                 self.status = "ERROR"
                 time.sleep(configs["SLEEP_TIME"])
@@ -223,10 +224,10 @@ Please enter your request to be completed🛸: """.format(art=text2art("UFO")), 
             self.cost += cost
 
             try:
-                try:
+                aad = configs['API_TYPE'].lower() == 'azure_ad'
+                if not aad:
                     response_string = response["choices"][0]["message"]["content"]
-                except:
-                    # response_string = response["choices"][0]["message"]["content"]
+                else:
                     response_string = response.choices[0].message.content
                 response_json = json_parser(response_string)
